feat(marketplace): Stripe-hosted Connect onboarding for creator payouts
Replaces the hand-rolled KYC form on /marketplace/creator/payout-setup
with Stripe's hosted onboarding via Account Links, and self-heals
creator profiles left without a Connect account (the root cause of the
400 on POST /api/marketplace/creator/payout-details).

- New service: createOnboardingLink wraps accountLinks.create with a
  mock-mode shortcut (returns returnUrl when STRIPE_SECRET_KEY unset)
- New route: POST /creator/payout-onboarding-link lazy-creates the
  Connect account if missing, then returns the Stripe-hosted URL
- /creator/payout-details now lazy-creates the Connect account too,
  unblocking any existing orphaned profiles without a redeploy
- payout-setup.tsx replaced with status pill + "Set up payouts with
  Stripe" CTA + currentlyDue field list; preserves ?legacy=1 escape
  hatch that renders the original form (extracted to
  components/marketplace/PayoutLegacyForm.tsx) so we can fall back
  without a redeploy if the hosted flow misbehaves
- Adds 4 unit tests covering throw/mock/live branches of
  createOnboardingLink (41/41 stripe-connect.service tests pass)
